+++ base-requirements.txt
@@ -1,4 +1,3 @@
 # VCS
 gt4py-functional @ git+https://github.com/GridTools/gt4py.git@572b04caaf4932a23ac0a6a266faa9ecb228807b
 
-cffi>=1.5
