+++ py2f/requirements-dev.txt
@@ -1,0 +1,4 @@
+-r ../base-requirements-dev.txt
+-e ../common
+-e ../atm_dyn_iconam
+-e .
