+++ py2f/requirements.txt
@@ -1,0 +1,4 @@
+-r ../base-requirements.txt
+../common
+../atm_dyn_iconam
+.
