+++ py2f/setup.cfg
@@ -1,0 +1,58 @@
+# This file is mainly used to configure package creation with setuptools.
+# Documentation:
+# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+#
+[metadata]
+name = icon4py_py2f
+description = Icon inspired code in Python and GT4Py
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/C2SM/icon4py
+author = ETH Zurich
+author_email = gridtools@cscs.ch
+license = gpl3
+license_files = LICENSE
+platforms = Linux, Mac
+classifiers =
+    Development Status :: 3 - Alpha
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)
+    Operating System :: POSIX
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: Implementation :: CPython
+    Topic :: Scientific/Engineering :: Atmospheric Science
+    Topic :: Scientific/Engineering :: Mathematics
+    Topic :: Scientific/Engineering :: Physics
+project_urls =
+    Source Code = https://github.com/C2SM/icon4py
+
+[options]
+packages = find_namespace:
+install_requires =
+    icon4py-common
+    icon4py-atm_dyn_iconam
+    cffi>=1.5
+    tabulate>=0.8.9
+    fprettify>=0.3.7
+python_requires = >=3.10
+package_dir =
+    = src
+zip_safe = False
+
+[options.package_data]
+# References:
+#  https://setuptools.pypa.io/en/latest/userguide/datafiles.html
+#  https://github.com/abravalheri/experiment-setuptools-package-data
+* = *.md, *.rst, *.toml, *.txt, py.typed
+
+[options.packages.find]
+where = src
+exclude =
+    tests
+
+[options.entry_points]
+console_scripts =
+    py2fgen = icon4py.py2f.wrapper.py2fgen:main
