+++ py2f/tests/test_code_generation.py
@@ -15,7 +15,7 @@ import string
 import pytest
 from functional.type_system.type_specifications import ScalarKind
 
-from icon4py.diffusion.wrapper.code_generation import (
+from icon4py.py2f.codegen import (
     CffiPlugin,
     CHeaderGenerator,
     DimensionType,
